Restructure Fresco for inclusion in amuse.ext
Work for issue #19
