+++ src/amuse/ext/fresco/__init__.py
@@ -1,0 +1,4 @@
+"""
+fresco generates images that resemble true observations.
+"""
+from amuse.ext.fresco.fresco import make_image
